services/auth/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
